separate tests & compile-time logging
